backend/requirements.txt
# prod
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.5
google-api-core==2.25.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# dev
pytest==8.3.3
ruff==0.6.9